# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Aerospike Backup Service",
        "title": "Backup Service REST API Specification",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Returns the configuration for the service.",
                "operationId": "readConfig",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Updates the configuration for the service.",
                "operationId": "updateConfig",
                "parameters": [
                    {
                        "description": "config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/clusters": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads all Aerospike clusters from the configuration.",
                "operationId": "readAllClusters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.AerospikeCluster"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/clusters/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads a specific Aerospike cluster from the configuration given its name.",
                "operationId": "readCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the Aerospike cluster",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    },
                    "404": {
                        "description": "The specified cluster could not be found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Updates an existing Aerospike cluster in the configuration.",
                "operationId": "updateCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "aerospike cluster",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Adds an Aerospike cluster to the config.",
                "operationId": "addCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "cluster info",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Deletes a cluster from the configuration by name.",
                "operationId": "deleteCluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AerospikeCluster": {
            "description": "AerospikeCluster represents the configuration for an Aerospike cluster for backup.",
            "type": "object",
            "properties": {
                "conn-timeout": {
                    "description": "The connection timeout in milliseconds.",
                    "type": "integer"
                },
                "credentials": {
                    "description": "The authentication details to the Aerospike cluster.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Credentials"
                        }
                    ]
                },
                "label": {
                    "description": "The cluster name.",
                    "type": "string"
                },
                "seed-nodes": {
                    "description": "The seed nodes details.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SeedNode"
                    }
                },
                "tls": {
                    "description": "The cluster TLS configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TLS"
                        }
                    ]
                },
                "use-services-alternate": {
                    "description": "Whether should use \"services-alternate\" instead of \"services\" in info request during cluster tending.",
                    "type": "boolean"
                }
            }
        },
...
        "model.BackupPolicy": {
            "description": "BackupPolicy represents a scheduled backup policy.",
            "type": "object",
            "properties": {
                "bandwidth": {
                    "description": "Throttles backup write operations to the backup file(s) to not exceed the given\nbandwidth in MiB/s.",
                    "type": "integer"
                },
                "file-limit": {
                    "description": "File size limit (in MB) for the backup directory. If an .asb backup file crosses this size threshold,\na new backup file will be created.",
                    "type": "integer"
                },
                "filter-exp": {
                    "type": "string"
                },
                "max-records": {
                    "description": "An approximate limit for the number of records to process. Available in server 4.9 and above.",
                    "type": "integer"
                },
                "max-retries": {
                    "description": "Maximum number of retries before aborting the current transaction.",
                    "type": "integer"
                },
                "no-bins": {
                    "description": "Only backup record metadata (digest, TTL, generation count, key).",
                    "type": "boolean"
                },
                "no-indexes": {
                    "description": "Do not back up any secondary index definitions.",
                    "type": "boolean"
                },
                "no-records": {
                    "description": "Do not back up any record data (metadata or bin data).",
                    "type": "boolean"
                },
                "no-udfs": {
                    "description": "Do not back up any UDF modules.",
                    "type": "boolean"
                },
                "parallel": {
                    "description": "Maximum number of scan calls to run in parallel.",
                    "type": "integer"
                },
                "records-per-second": {
                    "description": "Limit total returned records per second (RPS). If RPS is zero (the default),\nthe records-per-second limit is not applied.",
                    "type": "integer"
                },
                "remove-artifacts": {
                    "description": "Clear directory or remove output file.",
                    "type": "boolean"
                },
                "remove-files": {
                    "description": "Whether to clear the output directory (default: KeepAll).",
                    "enum": [
                        "KeepAll",
                        "RemoveAll",
                        "RemoveIncremental"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.RemoveFilesType"
                        }
                    ]
                },
                "retry-delay": {
                    "description": "RetryDelay defines the delay in milliseconds before retrying a failed operation.",
                    "type": "integer"
                },
                "socket-timeout": {
                    "description": "Socket timeout in milliseconds. If this value is 0, it is set to total-timeout.\nIf both are 0, there is no socket idle time limit.",
                    "type": "integer"
                },
                "total-timeout": {
                    "description": "Total socket timeout in milliseconds. Default is 0, that is, no timeout.",
                    "type": "integer"
                }
            }
        },
        "model.BackupRoutine": {
            "description": "BackupRoutine represents a scheduled backup operation routine.",
            "type": "object",
            "properties": {
                "after-digest": {
                    "type": "string"
                },
                "backup-policy": {
                    "description": "The name of the corresponding bakup policy.",
                    "type": "string"
                },
                "bin-list": {
                    "description": "The list of backup bin names (optional, an empty list implies backing up all bins).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "incr-interval-cron": {
                    "description": "The interval for incremental backup as a cron expression string (optional).",
                    "type": "string"
                },
                "interval-cron": {
                    "description": "The interval for full backup as a cron expression string.",
                    "type": "string"
                },
                "namespaces": {
                    "description": "The list of the namespaces to back up (optional, empty list implies backup up whole cluster).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "node-list": {
                    "description": "The list of nodes in the Aerospike cluster to run the backup for.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Node"
                    }
                },
                "partition-list": {
                    "description": "Back up list of partition filters. Partition filters can be ranges, individual partitions,\nor records after a specific digest within a single partition.\nDefault number of partitions to back up: 0 to 4095: all partitions.",
                    "type": "string"
                },
                "prefer-racks": {
                    "description": "A list of Aerospike Server rack IDs to prefer when reading records for a backup.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "secret-agent": {
                    "description": "The Secret Agent configuration for the routine (optional).",
                    "type": "string"
                },
                "set-list": {
                    "description": "The list of backup set names (optional, an empty list implies backing up all sets).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source-cluster": {
                    "description": "The name of the corresponding source cluster.",
                    "type": "string"
                },
                "storage": {
                    "description": "The name of the corresponding storage provider configuration.",
                    "type": "string"
                }
            }
        },
        "model.BackupServiceConfig": {
            "description": "BackupServiceConfig represents the backup service configuration properties.",
            "type": "object",
            "properties": {
                "http": {
                    "description": "HTTPServer is the backup service HTTP server configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.HTTPServerConfig"
                        }
                    ]
                },
                "logger": {
                    "description": "Logger is the backup service logger configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.LoggerConfig"
                        }
                    ]
                }
            }
        },
        "model.Config": {
            "description": "Config represents the service configuration file.",
            "type": "object",
            "properties": {
                "aerospike-clusters": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.AerospikeCluster"
                    }
                },
                "backup-policies": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.BackupPolicy"
                    }
                },
                "backup-routines": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.BackupRoutine"
                    }
                },
                "secret-agent": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.SecretAgent"
                    }
                },
                "service": {
                    "$ref": "#/definitions/model.BackupServiceConfig"
                },
                "storage": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.Storage"
                    }
                }
            }
        },
        "model.Credentials": {
            "description": "Credentials represents authentication details to the Aerospike cluster.",
            "type": "object",
            "properties": {
                "auth-mode": {
                    "description": "The authentication mode string (INTERNAL, EXTERNAL, EXTERNAL_INSECURE, PKI).",
                    "type": "string"
                },
                "password": {
                    "description": "The password for the cluster authentication.",
                    "type": "string"
                },
                "password-path": {
                    "description": "The file path with the password string, will take precedence over the password field.",
                    "type": "string"
                },
                "user": {
                    "description": "The username for the cluster authentication.",
                    "type": "string"
                }
            }
        },
...
        "model.HTTPServerConfig": {
            "description": "HTTPServerConfig represents the service's HTTP server configuration.",
            "type": "object",
            "properties": {
                "address": {
                    "description": "The address to listen on.",
                    "type": "string",
                    "default": "0.0.0.0"
                },
                "context-path": {
                    "description": "ContextPath customizes path for the API endpoints.",
                    "type": "string",
                    "default": "/"
                },
                "port": {
                    "description": "The port to listen on.",
                    "type": "integer",
                    "default": 8080
                },
                "rate": {
                    "description": "HTTP rate limiter configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.RateLimiterConfig"
                        }
                    ]
                }
            }
        },
...
        "model.LoggerConfig": {
            "description": "LoggerConfig represents the backup service logger configuration.",
            "type": "object",
            "properties": {
                "file-writer": {
                    "description": "File writer logging configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.FileLoggerConfig"
                        }
                    ]
                },
                "format": {
                    "description": "Format is the logger format (PLAIN, JSON).",
                    "type": "string",
                    "default": "PLAIN"
                },
                "level": {
                    "description": "Level is the logger level.",
                    "type": "string",
                    "default": "DEBUG"
                },
                "stdout-writer": {
                    "description": "Whether to enable logging to the standard output.",
                    "type": "boolean",
                    "default": true
                }
            }
        },
        "model.Node": {
            "description": "Node represents the Aerospike node details.",
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
...
        "model.RemoveFilesType": {
            "description": "RemoveFilesType represents the type of the backup storage.",
            "type": "string",
            "enum": [
                "KeepAll",
                "RemoveAll",
                "RemoveIncremental"
            ],
            "x-enum-varnames": [
                "KeepAll",
                "RemoveAll",
                "RemoveIncremental"
            ]
        },
...
        "model.SecretAgent": {
            "description": "SecretAgent represents the configuration of an Aerospike Secret Agent for a backup/restore operation.",
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address of the Secret Agent.",
                    "type": "string"
                },
                "port": {
                    "description": "Port the Secret Agent is running on.",
                    "type": "string"
                },
                "timeout": {
                    "description": "Timeout in milliseconds.",
                    "type": "integer"
                },
                "tls-ca": {
                    "description": "The path to a trusted CA certificate file in PEM format.",
                    "type": "string"
                },
                "tls-enabled": {
                    "description": "Indicates whether TLS is enabled.",
                    "type": "boolean"
                }
            }
        },
        "model.SeedNode": {
            "description": "SeedNode represents details of a node in the Aerospike cluster.",
            "type": "object",
            "properties": {
                "host-name": {
                    "description": "The host name of the node.",
                    "type": "string"
                },
                "port": {
                    "description": "The port of the node.",
                    "type": "integer"
                },
                "tls-name": {
                    "description": "TLS certificate name used for secure connections (if enabled).",
                    "type": "string"
                }
            }
        },
        "model.Storage": {
            "description": "Storage represents the configuration for a backup storage details.",
            "type": "object",
            "properties": {
                "path": {
                    "description": "The root path for the backup repository.",
                    "type": "string"
                },
                "s3-endpoint-override": {
                    "description": "An alternative endpoint for the S3 SDK to communicate (AWS S3 optional).",
                    "type": "string"
                },
                "s3-log-level": {
                    "description": "The log level of the AWS S3 SDK (AWS S3 optional).",
                    "type": "string",
                    "default": "Fatal"
                },
                "s3-profile": {
                    "description": "The S3 profile name (AWS S3 optional).",
                    "type": "string"
                },
                "s3-region": {
                    "description": "The S3 region string (AWS S3 optional).",
                    "type": "string"
                },
                "type": {
                    "description": "The type of the storage provider (0 - Local, 1 - AWS S3).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.StorageType"
                        }
                    ]
                }
            }
        },
        "model.StorageType": {
            "description": "StorageType represents the type of the backup storage.",
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "Local",
                "S3"
            ]
        },
        "model.TLS": {
            "description": "TLS represents the Aerospike cluster TLS configuration options.",
            "type": "object",
            "properties": {
                "cafile": {
                    "description": "Path to a trusted CA certificate file.",
                    "type": "string"
                },
                "capath": {
                    "description": "Path to a directory of trusted CA certificates.",
                    "type": "string"
                },
                "certfile": {
                    "description": "Path to the chain file for mutual authentication (if Aerospike Cluster supports it).",
                    "type": "string"
                },
                "cipher-suite": {
                    "description": "TLS cipher selection criteria. The format is the same as OpenSSL's Cipher List Format.",
                    "type": "string"
                },
                "keyfile": {
                    "description": "Path to the key for mutual authentication (if Aerospike cluster supports it).",
                    "type": "string"
                },
                "keyfile-password": {
                    "description": "Password to load protected TLS-keyfile (env:VAR, file:PATH, PASSWORD).",
                    "type": "string"
                },
                "name": {
                    "description": "The default TLS name used to authenticate each TLS socket connection.",
                    "type": "string"
                },
                "protocols": {
                    "description": "TLS protocol selection criteria. This format is the same as Apache's SSL Protocol.",
                    "type": "string"
                }
            }
        }
    },
...
}